HADOOP-19354. Factory init tuning.
StreamFactoryRequirements takes a list of requirement flags, and
moves off boolean flags. (incompatible change).

A new flag indicates that the stream may issue requests outside
an audit span. This is used when initializing the audit manager
so out-of-span requests are never rejected.

The requirement on vector IO is removed, instead set the #of
required threads in VectoredIOContext to zero (may revisit this)

VectoredIOContext moved to streams package and build() operation
makes immutable

* factory.bind() can throw an IOE

Change-Id: If2e121e6d3c0c6d19ac6f4b8452752a8bfafd3d1