HADOOP-19354. Stream factory and auditing flags; build option passdown
InputStreamFactory can return a set of flags after
initialization, these are used by S3AFileSystem to tune
behaviour in itself (thread pool options) *and* to disable
auditor rejection of out of span operations.

The out of span change is quite complicated as there's
a loop in the build.

auditor -> request factory -> store -> stream factory -> requirements

To address this there's an Auditor.setAuditFlags() option now.
This is not tested, though it will be once the analytics stream is wired
up.

Build:
It is nominally possible to set a build factory through maven

-Dstream=prefetch

However, this isn't being picked up as can be seen with runs of
-Dstream=custom
-Dstream=unknown

There MUST fail. they currently don't, except for a few test cases.

More work there needed

Change-Id: I76dc4782fdd1850f220368e4a394e1cfbc65adb9